requirements.txt
pygame>=2.0.0
PyOpenGL>=3.1.5
PyOpenGL_accelerate>=3.1.5
numpy>=1.19.0
